requirements.txt
keras
tensorflow
spotipy
opencv-python
pandas
numpy
matplotlib
flask
python-dotenv   